+++ src/components/error-boundary.jsx
@@ -1,7 +1,7 @@
 import React, { Component } from "react";
 import PropTypes from "prop-types";
-import logger from "./logger/logger";
-import fallback from "./assets/fallback.png";
+import logger from "../logger/logger";
+import fallback from "../assets/fallback.png";
 import "./error-boundary.scss";
 
 class ErrorBoundary extends Component {