+++ src/components/screen.jsx
@@ -3,10 +3,10 @@ import PropTypes from "prop-types";
 import SunCalc from "suncalc";
 import { createGrid } from "os2display-grid-generator";
 import Region from "./region";
-import logger from "./logger/logger";
-import TouchRegion from "./touch-region";
-import ConfigLoader from "./config-loader";
 import "./screen.scss";
+import logger from "../logger/logger";
+import TouchRegion from "./touch-region";
+import ConfigLoader from "../util/config-loader";
 
 /**
  * Screen component.