+++ src/components/touch-region.jsx
@@ -4,9 +4,9 @@ import "./touch-region.scss";
 import { createGridArea } from "os2display-grid-generator";
 import Slide from "./slide";
 import ErrorBoundary from "./error-boundary";
-import idFromPath from "./id-from-path";
-import IconClose from "./assets/icon-close.svg";
-import IconPointer from "./assets/icon-pointer.svg";
+import idFromPath from "../util/id-from-path";
+import IconClose from "../assets/icon-close.svg";
+import IconPointer from "../assets/icon-pointer.svg";
 
 /**
  * Region component.
@@ -35,7 +35,7 @@ function TouchRegion({ region }) {
    *
    * @param {object} slide - The slide.
    */
-  function slideDone(slide) {
+  const slideDone = (slide) => {
     setDisplayClose(false);
     setCurrentSlide(null);
 
@@ -47,7 +47,7 @@ function TouchRegion({ region }) {
       },
     });
     document.dispatchEvent(slideDoneEvent);
-  }
+  };
 
   /**
    * Handle region content event.